tests/test_requirements.txt
autopep8
flake8
coverage
coverage-badge
nose2
pytest-cov
